Moved Converter, Verifier and their respective tests to base cli package.
