+++ src/main/java/org/apache/commons/cli/Converter.java
@@ -14,7 +14,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package org.apache.commons.cli.converters;
+package org.apache.commons.cli;
 
 import java.io.File;
 import java.net.URL;
