+++ src/main/java/org/apache/commons/cli/Verifier.java
@@ -14,7 +14,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package org.apache.commons.cli.converters;
+package org.apache.commons.cli;
 
 import java.util.Arrays;
 import java.util.List;
