+++ src/test/java/org/apache/commons/cli/OptionTest.java
@@ -31,8 +31,6 @@ import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 
-import org.apache.commons.cli.converters.Converter;
-import org.apache.commons.cli.converters.Verifier;
 import org.junit.Test;
 
 public class OptionTest {
