+++ src/test/java/org/apache/commons/cli/TypeHandlerTest.java
@@ -32,8 +32,6 @@ import java.util.Date;
 import java.util.List;
 import java.util.stream.Stream;
 
-import org.apache.commons.cli.converters.Converter;
-import org.apache.commons.cli.converters.Verifier;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
